+++ lib/widgets/about/translators.dart
@@ -40,6 +40,7 @@ class AboutTranslators extends StatelessWidget {
     Contributor('Linerly', 'linerly@protonmail.com'),
     Contributor('Olexandr Mazur', 'rozihrash.ya6w7@simplelogin.com'),
     Contributor('vesp', 'vesp@post.cz'),
+    Contributor('Dan', 'denqwerta@gmail.com'),
     // Contributor('SAMIRAH AIL', 'samiratalzahrani@gmail.com'), // Arabic
     // Contributor('Salih Ail', 'rrrfff444@gmail.com'), // Arabic
     // Contributor('Piotr K', '1337.kelt@gmail.com'), // Polish