+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - Viewer: optionally show description on overlay
+- Collection: unlocated/untagged overlay icons
 - Czech translation (thanks vesp)
 
 ### Changed